final code for san mateo comp
