+++ TeamCode/build.gradle
@@ -22,6 +22,9 @@ android {
     }
 
     packagingOptions {
+        jniLibs {
+            pickFirsts += ['**/*.so']
+        }
         jniLibs.useLegacyPackaging true
     }
 }
