+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/ArmTest.java
@@ -8,6 +8,13 @@ import com.qualcomm.robotcore.util.ElapsedTime;
 @TeleOp(name="ArmTest")
 public class ArmTest extends LinearOpMode {
 
+//  static final double LEFT_ARM_BOTTOM_POSITION = 0.130;
+//    static final double RIGHT_ARM_BOTTOM_POSITION = 0.840;
+    static final double LEFT_ARM_BOTTOM_POSITION = 0.150;
+    static final double RIGHT_ARM_BOTTOM_POSITION = 0.785;
+    static final double RIGHT_ARM_DIFF = 0.740;
+    static final double LEFT_ARM_DIFF = RIGHT_ARM_DIFF + 0.005  ;
+
     // motors for claw
     private Servo leftArm = null;
     private Servo rightArm = null;
@@ -32,13 +39,13 @@ public class ArmTest extends LinearOpMode {
         while (opModeIsActive()) {
             boolean finger = false;
             if(gamepad1.a) {
-                rightArm.setPosition(0);
-                leftArm.setPosition(0.985);
+                leftArm.setPosition(LEFT_ARM_BOTTOM_POSITION);
+                rightArm.setPosition(RIGHT_ARM_BOTTOM_POSITION);
 
             }
             if(gamepad1.b) {
-                rightArm.setPosition(0.75);
-                leftArm.setPosition(0.23);
+                leftArm.setPosition(LEFT_ARM_BOTTOM_POSITION + LEFT_ARM_DIFF);
+                rightArm.setPosition(RIGHT_ARM_BOTTOM_POSITION - RIGHT_ARM_DIFF);
 
             }
             telemetry.update();
