+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/EruditeUtils.java
@@ -145,6 +145,30 @@ public class EruditeUtils {
         }
         while (motors[0].isBusy() || motors[1].isBusy() || motors[2].isBusy() || motors[3].isBusy());
     }
+    public void encoderDriveBackwardInaccurate(double inches, DcMotor motors[]) {
+
+        double     COUNTS_PER_MOTOR_REV    = 537.7;    // goBILDA Motor Encoder
+        double     DRIVE_GEAR_REDUCTION    = 1.0 ;     // No External Gearing.
+        double     WHEEL_DIAMETER_INCHES   = 3.77953 ;     // For figuring circumference
+        double     COUNTS_PER_INCH         = (COUNTS_PER_MOTOR_REV * DRIVE_GEAR_REDUCTION) / (WHEEL_DIAMETER_INCHES * 3.14159);
+        int [] newTarget = {0,0, 0, 0};
+
+
+        // Determine new target position, and pass to motor controller
+        newTarget[0] = motors[0].getCurrentPosition() - (int)(inches * COUNTS_PER_INCH);
+        newTarget[1] = motors[1].getCurrentPosition() - (int)(inches * COUNTS_PER_INCH);
+        newTarget[2] = motors[2].getCurrentPosition() - (int)(inches * COUNTS_PER_INCH);
+        newTarget[3] = motors[3].getCurrentPosition() - (int)(inches * COUNTS_PER_INCH);
+
+        for (int i = 0; i < 4; i+=1){
+            motors[i].setTargetPosition(newTarget[i]);
+        }
+
+        for (int i = 0; i < 4; i+=1){
+            motors[i].setMode(DcMotor.RunMode.RUN_TO_POSITION);
+        }
+        while (motors[0].isBusy() && motors[1].isBusy() && motors[2].isBusy() && motors[3].isBusy());
+    }
     public void encoderDriveRight(double inches, DcMotor motors[]) {
 
         double     COUNTS_PER_MOTOR_REV    = 537.7 ;    // goBILDA Motor Encoder
